feat(inventory): support decimal quantities
- Update database schemas and application logic to use decimal(12,3) for quantities.
- Add fractional quantity support to invoice and inventory modules.
- Extend trial duration to 100 years.
- Automate XAMPP file deployment in build script.

+++ db_sync_bridge.php
@@ -282,7 +282,7 @@ if ($_SERVER['REQUEST_METHOD'] === 'GET' && $action === 'sync_down') {
                 "typeName" => $item['typeName'],
                 "gradeName" => $item['gradeName'],
                 "unitName" => $item['unitName'],
-                "qty" => intval($item['qty']),
+                "qty" => floatval($item['qty']),
                 "price" => floatval($item['price']),
                 "total" => floatval($item['total']),
                 "priceType" => $item['priceType']
